added stuff, going to implement multithreading
