+++ ArcnumCore/core/window/window.hpp
@@ -5,8 +5,10 @@
 #include <glfw3.h>
 
 #include "renderer.hpp"
-#include "../world/voxel.hpp"
 #include "camera.hpp"
+#include "../world/voxel.hpp"
+#include "../world/voxel_manager.hpp"
+#include "../player/player.hpp"
 
 namespace arcnum_core
 {
@@ -16,7 +18,7 @@ namespace arcnum_core
 		window();
 		~window();
 		void init_gl();
-		void update(voxel_manager* entities, camera* player_camera, std::vector<world_position> voxel_positions);
+		void update(voxel_manager* entities, player* player, std::vector<world_position> voxel_positions);
 		bool is_running();
 		void handle_input(camera* player_camera);
 	public:
