+++ src/game/arcnum.hpp
@@ -3,6 +3,7 @@
 
 #include <core/window/window.hpp>
 #include <core/world/voxel.hpp>
+#include <core/player/player.hpp>
 
 namespace arcnum_main
 {
@@ -13,7 +14,7 @@ namespace arcnum_main
 		~arcnum();
 
 		void arcnum_main();
-		void main_loop(arcnum_core::camera* player_camera);
+		void main_loop(arcnum_core::player* player);
 	private:
 		arcnum_core::window* _main_window;
 		arcnum_core::voxel_manager* _main_entities;
