--- src/game/player/player.cpp
@@ -1,5 +1,0 @@
-#include "player.hpp"
-
-namespace arcnum_main
-{
-}
